#8 [REFACTOR] components/signup 폴더 리팩토링
